ai-apps-collection/agno-apps/ai-storyboard-generator/backend/requirements.txt
fastapi
uvicorn[standard]
agno
openai
python-dotenv
pillow
